requirements.txt
appengine-python-standard~=1.1
gql~=3.4
oauthlib~=3.2
pandas>=2.0
python-json-logger~=2.0
requests-oauthlib~=1.3
requests-toolbelt~=0.10
webcolors~=1.3
